// src/app/Models/userRegDto.ts
export class UserRegDto {
  firstname: string;

  lastname: string;

  password: string;

  email: string;

  phoneNo: string;

  companyName: string;

  companyAddress: string;

  designation: string;

  pinCode: string;

  bankId: number;

  account_number: string;

  ifsc_code: string;

  cardTypeId: number;

  addressLine1: string;

  addressLine2: string;

  city: string;

  panCardImg: any;

  adharCardImg: any;
}